refactor: deno_fetch op crate (#7524)
