+++ cli/rt/28_filereader.js
