+++ cli/tests/performance_stats.out
@@ -1,5 +1,5 @@
 [WILDCARD]
-Files: 45
+Files: 46
 Nodes: [WILDCARD]
 Identifiers: [WILDCARD]
 Symbols: [WILDCARD]
